requirements.txt
cerebras_cloud_sdk
exa-py
python-dotenv
embedchain
streamlit
requests
pydantic
httpx
uvicorn